Update upload.sh: Implement UPLOAD_WORKAROUND
A user had a wacky remote server that moved files he uploaded to a different directory, so checking for their existence failed.
By adding UPLOAD_WORKAROUND="true  to userVariables.sh the user was able to successfully upload files.

+++ scripts/upload.sh
@@ -388,28 +388,38 @@ else # sftp/ftp/ftps
 
 		# Unlikely, but just in case it's already there, remove it and log the action.
 		# Glob requires a regular expression, otherwise it always finds the file even if it's not there.
-		echo "glob --exist '[${TEMP_NAME:0:1}]${TEMP_NAME:1}' && echo '${TEMP_NAME} exists; removing...' && rm '${TEMP_NAME}'"
+		if [[ ${UPLOAD_WORKAROUND:-false} == "true" ]]; then
+			# Some web servers move the file just uploaded somewhere else so we can't rename it.
+			# So, upload the file with the final name.
+			# This means it may have problems overwritting the file if it's in use.
+			echo "put '${FILE_TO_UPLOAD}'
+				|| (echo 'put of ${FILE_TO_UPLOAD} failed!  Trying again...'; sleep 3;  put '${FILE_TO_UPLOAD}' && echo 'WORKED' && exit 0)
+				|| (echo '2nd put failed again, quitting'; exit 1)
+				|| exit 2"
+		else
+			echo "glob --exist '[${TEMP_NAME:0:1}]${TEMP_NAME:1}' && echo '${TEMP_NAME} exists; removing...' && rm '${TEMP_NAME}'"
 
-		echo "put '${FILE_TO_UPLOAD}' -o '${TEMP_NAME}'
-			|| (echo 'put of ${FILE_TO_UPLOAD} to ${TEMP_NAME} failed!  Trying again...'; sleep 3;  put '${FILE_TO_UPLOAD}' -o '${TEMP_NAME}' && echo 'WORKED' && exit 0)
-			|| (echo '2nd put failed again, quitting'; exit 1)
-			|| exit 2"
+			echo "put '${FILE_TO_UPLOAD}' -o '${TEMP_NAME}'
+				|| (echo 'put of ${FILE_TO_UPLOAD} to ${TEMP_NAME} failed!  Trying again...'; sleep 3;  put '${FILE_TO_UPLOAD}' -o '${TEMP_NAME}' && echo 'WORKED' && exit 0)
+				|| (echo '2nd put failed again, quitting'; exit 1)
+				|| exit 2"
 
-		# Try to remove ${DESTINATION_NAME}, which may or may not exist.
-		# If the "rm" fails, the file may be in use by the web server or another lftp,
-		# so wait a few seconds and try again, but without the "-f" option so we see any error msg.
-		echo "rm  -f '${DESTINATION_NAME}'"
-		echo "glob --exist '[${DESTINATION_NAME:0:1}]${DESTINATION_NAME:1}'
-			&& (echo 'rm of ${DESTINATION_NAME} failed!  Trying again...';  sleep 3;  rm '${DESTINATION_NAME}' && echo 'WORKED' && exit 1)
-			&& (echo '2nd rm failed, quiting.'; rm -f '${TEMP_NAME}'; exit 1)
-			&& exit 3"
+			# Try to remove ${DESTINATION_NAME}, which may or may not exist.
+			# If the "rm" fails, the file may be in use by the web server or another lftp,
+			# so wait a few seconds and try again, but without the "-f" option so we see any error msg.
+			echo "rm  -f '${DESTINATION_NAME}'"
+			echo "glob --exist '[${DESTINATION_NAME:0:1}]${DESTINATION_NAME:1}'
+				&& (echo 'rm of ${DESTINATION_NAME} failed!  Trying again...';  sleep 3;  rm '${DESTINATION_NAME}' && echo 'WORKED' && exit 1)
+				&& (echo '2nd rm failed, quiting.'; rm -f '${TEMP_NAME}'; exit 1)
+				&& exit 3"
 
-		# If the first "mv" fails, wait, then try again.  If that works, exit 0.
-		# If the 2nd "mv" also fails exit 4.  Either way, display a 2nd message.
-		echo "mv '${TEMP_NAME}' '${DESTINATION_NAME}'
-			|| (echo 'mv of ${TEMP_NAME} to ${DESTINATION_NAME} failed!  Trying again...'; sleep 3; mv '${TEMP_NAME}' '${DESTINATION_NAME}' && echo 'WORKED' && exit 0)
-			|| (echo '2nd mv failed, quitting.'; rm -f '${TEMP_NAME}'; exit 1)
-			|| exit 4"
+			# If the first "mv" fails, wait, then try again.  If that works, exit 0.
+			# If the 2nd "mv" also fails exit 4.  Either way, display a 2nd message.
+			echo "mv '${TEMP_NAME}' '${DESTINATION_NAME}'
+				|| (echo 'mv of ${TEMP_NAME} to ${DESTINATION_NAME} failed!  Trying again...'; sleep 3; mv '${TEMP_NAME}' '${DESTINATION_NAME}' && echo 'WORKED' && exit 0)
+				|| (echo '2nd mv failed, quitting.'; rm -f '${TEMP_NAME}'; exit 1)
+				|| exit 4"
+		fi
 
 		echo exit 0
 	} > "${LFTP_CMDS}"
